requirements.txt
torchsummary
absl-py==0.8.1
adabound==0.0.5
asn1crypto==1.0.1
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
conda==4.7.12
conda-build==3.18.9
conda-package-handling==1.6.0
cryptography==2.7
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
defusedxml==0.6.0
easydict==1.9
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gast==0.2.2
glob2==0.7
google-pasta==0.1.7
grpcio==1.16.1
h5py==2.9.0
idna==2.8
imageio==2.6.1
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
json5==0.8.5
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyter-tensorboard==0.1.10
jupyterlab==1.1.4
jupyterlab-server==1.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
libarchive-c==2.8
lief==0.9.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mkl-fft==1.0.14
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==7.2.0
msgpack==0.6.2
msgpack-numpy==0.4.4.3
natsort==6.2.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
notebook==6.0.2
numpy==1.16.4
olefile==0.46
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.2
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.0
pkginfo==1.5.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.9.2
psutil==5.6.3
ptyprocess==0.6.0
pycosat==0.6.3
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.4
pyrsistent==0.15.4
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
ruamel-yaml==0.15.46
scikit-image==0.16.2
scikit-learn==0.20.2
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.3
statsmodels==0.10.1
tabulate==0.8.6
tensorboard==2.0.0
tensorboardX==1.9
tensorflow==2.0.0
tensorflow-estimator==2.0.0
tensorpack==0.9.8
termcolor==1.1.0
terminado==0.8.2
test-tube==0.7.5
testpath==0.4.2
torch==1.3.0
torchvision==0.4.1a0+d94043a
tornado==6.0.3
tqdm==4.35.0
traitlets==4.3.3
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
kornia==0.2.0